+++ app/.babelrc.js
@@ -3,7 +3,6 @@ module.exports = {
     "@babel/preset-react",
   ],
   "plugins": [
-    "react-hot-loader/babel",
     "@babel/plugin-syntax-dynamic-import"
   ],
   "env": {
