+++ app/src/buy-section.js
