+++ app/src/position-group-details.js
