+++ package.json
@@ -67,7 +67,6 @@
     "lint-staged": "^8.1.5",
     "node-sass": "^4.12.0",
     "prettier": "^1.17.0",
-    "react-hot-loader": "^4.8.4",
     "rlp": "^2.2.3",
     "run-with-testrpc": "^0.3.1",
     "sass-loader": "^7.1.0",
@@ -77,7 +76,6 @@
     "webpack": "^4.30.0",
     "webpack-bundle-analyzer": "^3.3.2",
     "webpack-cli": "^3.3.1",
-    "webpack-common-shake": "^2.1.0",
     "webpack-dev-server": "^3.3.1"
   },
   "husky": {
